requirements.txt
Flask==2.3.2
google-generative-ai==0.3.2  # For Gemini API